backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
jinja2
google-generativeai
